+++ btgsolutions_dataservices/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "3.3.0"
+__version__ = "4.0.0"
 
 from .websocket import *
 from .rest import *